requirements.txt
torch
torchvision
transformers
pillow
numpy
matplotlib
tqdm
fastapi
uvicorn[standard]
python-multipart
aiofiles    